requirements.txt
python-telegram-bot
python-dateutil
python-dotenv
pymongo[srv]
validators
requests
